tests: sftp listdir under a locale with non-ascii date strings
This demonstrates failure under python-2.7, because SFTPAttributes
uses the locale-aware %b directive for the abbreviated month name
with time.strftime(), and it is not decoded correctly with python-2.7.

Instead of fixing for python-2.7 in paramiko-ng, we'll just wait
until python2 support is removed before merging in this test case.
This test code was so old, it needed multiple rounds of updates
and cleanups by multiple contributors:

Co-Authored-By: Jeff Forcier <jeff@bitprophet.org>
Co-Authored-By: Stanislav Levin <slev@altlinux.org>