improve performance by calling Message.asbytes() directly
Type-switch based code is used to convert a value to bytes.
However, in all cases (except for one, fixed in this PR),
the caller is specifying a Message instance. We can make
this code much simpler by directly calling the correct method on Message.

(This might not have as large an effect on this fork as on upstream,
because this fork doesn't use try/except for type check in asbytes() helper.)